Clearer instructions
The instructions on the readme is a little confusing,

For example if my bucket name was go_couch513 I ended up trying something like go_couch513.s3.amazonaws.com. 8 minutes later I realized it was just go_couch513 (Yes I know it is common for people to name their bucket with a .com) but this should make it more clearer for future developers.